python-ai/requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic
pytesseract
pdfplumber
Pillow